# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/settings": {
      "get": {
        "operationId": "SettingsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Settings"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Settings"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "description": "Requer um token válido de um usuário ainda não cadastrado (fluxo de cadastro incompleto).",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{user_id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_deactivate",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WattAPI",
    "description": "Documentação gerada a partir dos schemas zod de request/response e do metadata de @RoutePolicy de cada rota.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AppSettingsDto": {
        "type": "object",
        "properties": {
          "min_week_hours": {
            "type": "number"
          },
          "min_availability_hours": {
            "type": "number"
          }
        },
        "required": [
          "min_week_hours",
          "min_availability_hours"
        ],
        "example": {
          "min_week_hours": 40,
          "min_availability_hours": 0
        }
      },
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {
          "min_week_hours": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 98,
            "minimum": 0
          },
          "min_availability_hours": {
            "type": "integer",
            "minimum": 0,
            "maximum": 98
          }
        },
        "example": {
          "min_week_hours": 44,
          "min_availability_hours": 10
        }
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "sector": {
            "type": "string",
            "enum": [
              "projetos",
              "comercial",
              "marketing",
              "executivo",
              "institucional"
            ]
          },
          "cpf": {
            "type": "string",
            "pattern": "^([0-9]{11}|[0-9]{3}\\.[0-9]{3}\\.[0-9]{3}-[0-9]{2})$"
          }
        },
        "required": [
          "name",
          "sector",
          "cpf"
        ],
        "example": {
          "name": "João Silva",
          "sector": "projetos",
          "cpf": "12345678901"
        }
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sector": {
            "type": "string"
          },
          "cpf": {
            "description": "Presente apenas se quem consulta tem rank >= 2 (gerente ou acima) ou é o próprio usuário consultado — ver RoutePolicy.output.",
            "type": "string"
          },
          "house_id": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "sector",
          "house_id",
          "created_at",
          "updated_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "email": "joao@empresa.com",
          "name": "João Silva",
          "role": "consultor",
          "sector": "projetos",
          "cpf": "123.456.789-09",
          "house_id": null,
          "created_at": "2026-01-15T10:00:00.000Z",
          "updated_at": "2026-05-20T14:30:00.000Z"
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "enum": [
              "consultor",
              "gerente",
              "diretor",
              "presidente",
              "assessor"
            ]
          },
          "sector": {
            "type": "string",
            "enum": [
              "projetos",
              "comercial",
              "marketing",
              "executivo",
              "institucional"
            ]
          },
          "cpf": {
            "type": "string",
            "pattern": "^([0-9]{11}|[0-9]{3}\\.[0-9]{3}\\.[0-9]{3}-[0-9]{2})$"
          }
        },
        "example": {
          "email": "novo@empresa.com",
          "name": "Novo Nome",
          "role": "gerente",
          "sector": "comercial",
          "cpf": "98765432100"
        }
      },
...
    }
  }
}